Cleaning up schema alchemy stuff
